pyproject.toml
[project]
name = "alfred"
version = "0.2.0"
description = "API that manages wgid pipeline."
readme = "README.md"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.4",
    "Programming Language :: Python :: 3.5",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
requires-python = ">=3.4"
license = {file = "LICENSE"}
dependencies = [
    "pytest",
    "requests",
    "toml",
    "PyYAML",
    "confuse",
    "anytree",
    "requests",
    "bs4",
]
#[tool.setuptools]
## ...
#package-dir = {"" = "src"}

#[tool.setuptools.package-dir]
#mypkg = "src"

[tool.setuptools.packages.find]
where = ["src"]
include = ["alfred*"]  # alternatively: `exclude = ["additional*"]`
namespaces = false

[project.scripts]
wgid-shot-create = "alfred.cli.maker:create"
showtime = "alfred.showtime:its_showtime"
st = "alfred.showtime:its_showtime"

# Poety Setup
[tool.poetry]
name = "alfred"
version = "0.1.0"
description = "Database pipeline manager and creater"
authors = ["robert-v <robertvigorito@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<4"
PyYAML = "^5.4.1"
confuse = "^1.4.0"
anytree = "^2.8.0"
requests = "^2.25.1"
bs4 = "^0.0.1"
pymongo = "^4.3.3"
python-dotenv = "^1.0.0"

[tool.poetry.group.test.dependencies]
toml = "^0.10.2"
pytest = "^7.2.1"
pytest-cov = "^4.1.0"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 120
disable = "C0330, R0201, 0326"

[tool.tox]
legacy_tox_ini = """
[tox]
isolated_build = true
envlist = py37, py38, py39, py310

[testenv]
allowlist_externals = poetry,echo
commands =
    pip install -U pip
    pip install . pytest
    pytest -s -rap .
"""